requirements.txt
aiogram>=3.20.0
psutil>=7.0.0
python-dotenv>=1.1.0
